backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-dotenv
google-generativeai
pydantic